Split review into several actions + move some code around
